# fd-microservice/openapi_spec.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/report/fd-summary":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "report-controller"
                                                                      ],
                                                             "operationId":  "getFdSummary",
                                                             "responses":  {
                                                                               "404":  {
                                                                                           "description":  "Not Found",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "400":  {
                                                                                           "description":  "Bad Request",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "500":  {
                                                                                           "description":  "Internal Server Error",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "403":  {
                                                                                           "description":  "Forbidden",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "$ref":  "#/components/schemas/FdSummaryReport"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                             },
                  "/api/report/export/csv":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "report-controller"
                                                                      ],
                                                             "operationId":  "exportCsv",
                                                             "responses":  {
                                                                               "404":  {
                                                                                           "description":  "Not Found",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "400":  {
                                                                                           "description":  "Bad Request",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "500":  {
                                                                                           "description":  "Internal Server Error",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "403":  {
                                                                                           "description":  "Forbidden",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       },
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "type":  "string",
                                                                                                                                                    "format":  "byte"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         }
                                             },
                  "/api/report/customer-portfolio":  {
                                                         "get":  {
                                                                     "tags":  [
                                                                                  "report-controller"
                                                                              ],
                                                                     "operationId":  "getCustomerPortfolio",
                                                                     "responses":  {
                                                                                       "404":  {
                                                                                                   "description":  "Not Found",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               },
                                                                                       "400":  {
                                                                                                   "description":  "Bad Request",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               },
                                                                                       "500":  {
                                                                                                   "description":  "Internal Server Error",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               },
                                                                                       "403":  {
                                                                                                   "description":  "Forbidden",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               },
                                                                                       "200":  {
                                                                                                   "description":  "OK",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "type":  "array",
                                                                                                                                              "items":  {
                                                                                                                                                            "$ref":  "#/components/schemas/FdPortfolioReport"
                                                                                                                                                        }
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               }
                                                                                   }
                                                                 }
                                                     },
...
              },
    "components":  {
                       "schemas":  {
                                       "ApiResponse":  {
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "success":  {
                                                                                              "type":  "boolean"
                                                                                          },
                                                                              "message":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                              "data":  {
                                                                                           "type":  "object"
                                                                                       },
                                                                              "timestamp":  {
                                                                                                "type":  "string",
                                                                                                "format":  "date-time"
                                                                                            }
                                                                          }
                                                       },
...
                                       "FdSummaryReport":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "productCode":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                  "productName":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                  "totalAccounts":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int64"
                                                                                                    },
                                                                                  "totalPrincipal":  {
                                                                                                         "type":  "number"
                                                                                                     },
                                                                                  "totalInterestAccrued":  {
                                                                                                               "type":  "number"
                                                                                                           },
                                                                                  "activeAccounts":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int64"
                                                                                                     },
                                                                                  "closedAccounts":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int64"
                                                                                                     }
                                                                              }
                                                           },
                                       "FdPortfolioReport":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "fdAccountNo":  {
                                                                                                        "type":  "string"
                                                                                                    },
                                                                                    "productCode":  {
                                                                                                        "type":  "string"
                                                                                                    },
                                                                                    "principalAmount":  {
                                                                                                            "type":  "number"
                                                                                                        },
                                                                                    "interestRate":  {
                                                                                                         "type":  "number"
                                                                                                     },
                                                                                    "tenureMonths":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int32"
                                                                                                     },
                                                                                    "status":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                    "maturityDate":  {
                                                                                                         "type":  "string",
                                                                                                         "format":  "date"
                                                                                                     },
                                                                                    "accruedInterest":  {
                                                                                                            "type":  "number"
                                                                                                        },
                                                                                    "projectedMaturityAmount":  {
                                                                                                                    "type":  "number"
                                                                                                                }
                                                                                }
                                                             },
...
                                   }
                   }
}